+++ auth-web/src/components/auth/BcolLogin.vue
@@ -105,6 +105,9 @@ export default class BcolLogin extends Vue {
       } catch (err) {
         this.isLoading = false
         switch (err.response.status) {
+          case 409:
+            this.errorMessage = err.response.data.message
+            break
           case 400:
             this.errorMessage = err.response.data.message
             break
